requirements_upbitMA.txt
requests>=2.28.0
pykrx==1.0.48
pandas==2.2.3
openpyxl==3.1.5
psutil==7.0.0
exchange_calendars==4.10
python-dotenv>=1.0.0